requirements.txt
fastapi
uvicorn
python-dotenv
httpx
anthropic